Refactoring to move screens out of MainActivity.
