+++ app/src/main/java/com/jordangellatly/hermosahappyhour/DetailScreen.kt
@@ -1,0 +1,47 @@
+package com.jordangellatly.hermosahappyhour
+
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.*
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.ArrowBack
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.navigation.NavController
+import androidx.navigation.compose.rememberNavController
+
+@Composable
+fun DetailScreen(
+    navController: NavController
+) {
+    Scaffold(
+        topBar = {
+            TopAppBar(
+                title = { Text(text = "Detail View") },
+                navigationIcon = if (navController.previousBackStackEntry != null) {
+                    {
+                        IconButton(onClick = { navController.navigateUp() }) {
+                            Icon(
+                                imageVector = Icons.Filled.ArrowBack,
+                                contentDescription = "Back"
+                            )
+                        }
+                    }
+                } else {
+                    null
+                }
+            )
+        },
+        bottomBar = {
+            BottomNavigationBar(navController = navController)
+        }
+    ) { contentPadding ->
+        Text(text = "Hello", modifier = Modifier.padding(contentPadding))
+    }
+}
+
+@Preview
+@Composable
+fun DetailScreenPreview() {
+    DetailScreen(navController = rememberNavController())
+}
