+++ app/src/main/java/com/jordangellatly/hermosahappyhour/HomeScreen.kt
@@ -1,0 +1,138 @@
+package com.jordangellatly.hermosahappyhour
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.material.*
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.navigation.NavController
+import androidx.navigation.compose.rememberNavController
+import com.jordangellatly.hermosahappyhour.ui.theme.HermosaHappyHourTheme
+
+@Composable
+fun HomeScreen(
+    navController: NavController,
+    restaurants: List<Restaurant>
+) {
+    Scaffold(
+        topBar = {
+            TopAppBar(
+                backgroundColor = MaterialTheme.colors.primary,
+                title = { Text(stringResource(id = R.string.app_name)) }
+            )
+        },
+        bottomBar = {
+            BottomNavigationBar(navController = navController)
+        }
+    ) { contentPadding ->
+        LazyColumn(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(contentPadding),
+            verticalArrangement = Arrangement.spacedBy(16.dp),
+            contentPadding = PaddingValues(16.dp)
+        ) {
+            item {
+                Row(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .wrapContentHeight()
+                        .padding(vertical = 25.dp),
+                    horizontalArrangement = Arrangement.Center,
+                    verticalAlignment = Alignment.CenterVertically
+                ) {
+                    Text(
+                        text = "Restaurant List",
+                        style = MaterialTheme.typography.h3
+                    )
+                }
+            }
+            items(restaurants) { restaurant ->
+                RestaurantCard(
+                    restaurant.name,
+                    restaurant.description,
+                    restaurant.image,
+                    onClick = {
+                        navController.navigate("detail")
+                    }
+                )
+            }
+        }
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+fun HomeScreenPreview() {
+    HermosaHappyHourTheme {
+        HomeScreen(
+            navController = rememberNavController(),
+            restaurants = generateRestaurants()
+        )
+    }
+}
+
+@OptIn(ExperimentalMaterialApi::class)
+@Composable
+fun RestaurantCard(
+    name: String,
+    description: String,
+    image: Int,
+    onClick: () -> Unit
+) {
+    Card(
+        modifier = Modifier
+            .padding(10.dp)
+            .fillMaxWidth()
+            .wrapContentHeight(),
+        shape = MaterialTheme.shapes.medium,
+        elevation = 5.dp,
+        backgroundColor = MaterialTheme.colors.surface,
+        onClick = onClick
+    ) {
+        Row(
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Image(
+                painter = painterResource(id = image),
+                contentDescription = null,
+                modifier = Modifier
+                    .size(130.dp)
+                    .padding(8.dp),
+                contentScale = ContentScale.Fit,
+            )
+            Column(
+                modifier = Modifier.padding(8.dp)
+            ) {
+                Text(
+                    text = name,
+                    style = MaterialTheme.typography.h4,
+                    color = MaterialTheme.colors.onSurface
+                )
+                Text(
+                    text = description,
+                    style = MaterialTheme.typography.body2
+                )
+            }
+        }
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+fun RestaurantCardPreview() {
+    RestaurantCard(
+        name = "Sharkeez",
+        description = "Best margaritas in town!",
+        image = R.drawable.sharkeez,
+        onClick = {}
+    )
+}
